feat: rename tag → kind across entire codebase
Renamed the discriminated union tag field/enum from `tag` to `kind` everywhere:
emitters, analyzers, code fixes, tests, benchmarks, docs, and snapshot outputs.

Also includes benchmark suite updates (BenchmarkConstants, worst-case scenarios,
JSON property patterns) and AddMissingArmsCodeFix property pattern support.

+++ src/Spire.Analyzers/Rules/SPIRE015ExhaustiveEnumSwitchAnalyzer.cs
@@ -316,7 +316,7 @@ public sealed class SPIRE015ExhaustiveEnumSwitchAnalyzer : DiagnosticAnalyzer
             short s => s,
             ushort us => us,
             int i => i,
-            uint ui => (long)ui,
+            uint ui => ui,
             long l => l,
             ulong ul => (long)ul,
             _ => 0,
